Moved test configuration for functional test
