+++ src/main/java/uk/gov/hmcts/reform/jobscheduler/config/ServiceTokenConfiguration.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.jobscheduler.config;
 
-import org.apache.commons.lang.NotImplementedException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
@@ -10,10 +9,6 @@ import org.springframework.context.annotation.Lazy;
 import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-import uk.gov.hmcts.reform.authorisation.validators.AuthTokenValidator;
-import uk.gov.hmcts.reform.authorisation.validators.ServiceAuthTokenValidator;
-
-import java.util.List;
 
 @Configuration
 @Lazy
@@ -34,35 +29,4 @@ public class ServiceTokenConfiguration {
                 serviceAuthorisationApi
             );
     }
-
-    @Bean
-    @ConditionalOnProperty(name = "idam.s2s-auth.url", havingValue = "false")
-    public AuthTokenGenerator tokenGeneratorStub() {
-        return () -> "newly-generated-token";
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "idam.s2s-auth.url")
-    public AuthTokenValidator tokenValidator(ServiceAuthorisationApi s2sApi) {
-        return new ServiceAuthTokenValidator(s2sApi);
-    }
-
-    @Bean
-    @ConditionalOnProperty(name = "idam.s2s-auth.url", havingValue = "false")
-    public AuthTokenValidator tokenValidatorStub() {
-        return new AuthTokenValidator() {
-            public void validate(String token) {
-                throw new NotImplementedException();
-            }
-
-            public void validate(String token, List<String> roles) {
-                throw new NotImplementedException();
-            }
-
-            public String getServiceName(String token) {
-                return "some_service_name";
-            }
-        };
-    }
-
 }
